Add files via upload
Diversity and Inclusion (D&I) analysis is a process of examining and evaluating the representation, engagement, and treatment of individuals from various backgrounds within a specific context, such as an organization, community, or society. It involves collecting, assessing, and interpreting data to understand the level of diversity and the extent to which an inclusive environment is fostered. The goal of such analysis is to identify disparities, biases, and opportunities for improvement, and to develop strategies that promote equality, equity, and belonging for all individuals, regardless of their characteristics.